+++ sshd-common/src/test/java/org/apache/sshd/server/keyprovider/PEMGeneratorHostKeyProviderTest.java
@@ -118,11 +118,11 @@ public class PEMGeneratorHostKeyProviderTest extends JUnitTestSupport {
     }
 
     private static KeyPair validateKeyPairProvider(KeyPairProvider provider, String keyType) {
-        Iterable<String> types = provider.getKeyTypes();
+        Iterable<String> types = provider.getKeyTypes(null);
         KeyPair kp = null;
         for (String type : types) {
             if (keyType.equals(type)) {
-                kp = provider.loadKey(keyType);
+                kp = provider.loadKey(null, keyType);
                 assertNotNull("Failed to load key for " + keyType, kp);
                 break;
             }